backend/requirements.txt
fastapi
uvicorn
httpx

pydantic
pydantic-core
pydantic-settings
pydantic-extra-types

fastapi-users
fastapi-users-db-sqlalchemy
email-validator
python-multipart
bcrypt
passlib
pyjwt

sqlalchemy
asyncpg
alembic
mako

celery
redis

python-dotenv
typing-extensions